examples/coding_workforce/requirements.txt
# Install agents-hub in development mode
-e .

# Additional dependencies
python-dotenv>=1.0.0
